dockerfile/requirements.txt
fastapi
pydantic
mysql-connector-python
mysql
SQLAlchemy==1.4.39
python-dotenv
uvicorn==0.18.2
PyMySQL==1.0.2
line-bot-sdk==2.4.2
python-barcode==0.15.1
Pillow==9.0.1
python-multipart==0.0.6
azure-storage-blob==12.18.3